삭제를 눌렀을 때 비로그인 시 로그인하면GET 요청으로 가서 405에러가 뜨는 문제 해결

+++ articles/views.py
@@ -63,10 +63,10 @@ def detail(request, pk):
     return render(request, 'articles/detail.html', context)
 
 @require_POST
-@login_required
 def delete(request, pk):
-    article = get_object_or_404(Article,pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article,pk=pk)
+        article.delete()
     return redirect('articles:index')
